Fix: TileMap rendering problem
Add sound for witch in stage2
