--- Assets/Explode.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 42235ee7d6eb94f089f3f0c6a9af62a0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
