+++ Assets/Scenes/Main.unity
@@ -4398,6 +4398,7 @@ MonoBehaviour:
   - {fileID: 8300000, guid: a835f9cad9cb2412bbfc5212d080e0f6, type: 3}
   - {fileID: 8300000, guid: 6bd88a369fa664864bd3bda522e06386, type: 3}
   - {fileID: 8300000, guid: bfecec87b5dd44bfebfb2a44bcb23426, type: 3}
+  - {fileID: 8300000, guid: 1d72083ec6e0a4f978e93dea80dafef7, type: 3}
   sfxVolume: 0.4761583
 --- !u!4 &1181962955
 Transform:
