+++ Assets/Scripts/Tower/ShowTurretRange.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e4897c67a608e48ce96e5067a4744918
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
