+++ Assets/res/Map/Tileset Topdown - Monster Architect Overworld/0_Overworld/Overworld_Ground.png.meta
@@ -72,7 +72,7 @@ TextureImporter:
     maxTextureSize: 2048
     resizeAlgorithm: 0
     textureFormat: -1
-    textureCompression: 1
+    textureCompression: 2
     compressionQuality: 50
     crunchedCompression: 0
     allowsAlphaSplitting: 0
